django-heroku-docker/requirements.txt
gunicorn==20.1.0
django==5.0.1
whitenoise==6.6.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
django-heroku==0.3.1
dj-database-url==2.1.0


